refactor: split create creds comonent from dialog
